Restore material names to farlook
Item material names were removed from any use of 'distant_oname' in
bb90455822, intended to target a specific bug in cursed_gear_welds;
this made it impossible to determine the material of an item from a
distance, since farlook uses distant_oname.

Fix the bug which prompted bb90455822 in a slightly different way, by
copying over the material for the fake object in minimal_xname if it's
an item type which includes its default materials in its name.  This
means the cursed_gear_welds message will be 'the plastic gloves weld
themselves...' rather than 'the gloves weld themselves...' for objects
which force the inclusion of material names, but I think it's worth the
slight loss of minimalism for those particular item types to regain the
ability to judge object materials from a distance.