Fix: specifying material in wish for ring mail
A false positive match with '[material] ring' when wishing for
'[material] ring mail' meant that the wish wouldn't be recognized and
the player would get a 'nothing fitting that description exists' msg.
Add a specific check for 'ring mail' in not_actually_specifying_material
so that it won't produce false positives in this case.

Of course, it's impossible to wish for a particular material outside of
wizard mode anyway, so this is unlikely to impact a normal game.