Clarify feedback for aborted container transfer
...due to only locked containers being available

The message "there's no other container to put items in" suggests that
no other container existed on the spot (or in player inventory) at all,
but would also be printed if the only other containers available were
known to be locked.  Clarify this a bit by rephrasing the message and
specifying that no 'unlocked containers' exist, if the only ones
accessible are locked.