Clarify flint-making 'crumbles' message
'It crumbles' is ambiguous about which of the two items involved is
crumbling, so make the message more specific by naming the rock.

Also distinguish between a rock 'crumbling', producing flint and
destroying the rock, and 'crumbling slightly', which has no effect.
Although the commit message in b5a46e38 seems to suggest otherwise, the
rock is only destroyed when flint is produced: make it clear that the
rock isn't completely crumbling in cases where nothing happens.