Fix: invalid material on prayer gift spellbook
When generating a spellbook as a boon from prayer, the object is
generated, and then the otyp may be changed if it is an already-known
spell or some other undesireable spellbook type.  This could result in
the spellbook's material being mismatched, if the original spellbook was
vellum or parchment (which are leather) and the otyp was changed to
another spellbook (others are paper), triggering an 'invalid material'
sanity check.  Make sure to reset the spellbook material to default
after potentially re-rolling the otyp.