Restore ability to bang rocks together
Rock-breaking for flint added the ability to #rub two rocks together to
produce flint, but the callbacks for those actions added in the getobj
refactor didn't permit selecting rocks.  Restore the ability to choose
a rock, but don't suggest it (as was the case prior to the changes to
getobj, as far as I can see).