Fix: confused enchant weapon reset potion dilution
Confused enchant weapon was expanded to work on potions, rings, etc, but
because oeroded tracks dilution for potions, it was resetting potion
dilution in the process of making the potion 'indestructible'. Only
reset oeroded/2 for erosion_matters objects, since other objects may
use those bits for other purposes which don't make sense for fooproofing
to reset.