Reset crysknife material in hold_another_object
When adding an artifact, hold_another_object sets it on the floor,
then picks it back up if you survive touching it -- this process can
make a crysknife revert back to a worm tooth, even though it's not
'actually' being put onto the floor, so hold_another_object will set it
back to a crysknife before adding it to inventory.  Make sure the
material is reset as well as the otyp and oerodeproof status, since
reverting back to a worm tooth will reset the object's material to bone.

This is a latent bug since there isn't any artifact crysknife in the
game, but may as well make sure it's handled properly.