Fix: Gehennom maze rooms placed incorrectly
For mazewalk to work properly, random rooms in maze levels must be
placed so that the walls are on even coordinates (so that they line up
with the walls of the maze).  check_room was adjusting the positioning
and size of the rooms without adhering to those rules, leading to rooms
that were offset from where they should be.

This could also cause doors to show up in weird places, because door
placement in maze_add_rooms used the original lx,ly,hx,hy from before
the room size/position change by check_room.

I think this issue is related to #56; the leftmost room in the
screenshot has an invalid y-coordinate; that may be the cause of, or at
least a contributing factor to, the problem with mazewalk.