Add sanity checking for maze generation
Produces an impossible when a maze fails to cover the entire level.

Note by copperwater: This may end up obsolete in the Gehennom revamp
planned for 8.0, in which mazes are likely going away, but as long as
that doesn't end up attempting to call check_maze_coverage still, it
shouldn't be a problem.