Fix #56: failed mazewalk
walkfrom's recursion worked in such a way that it could go back to every
level _but_ the very first position it started from. This meant that if
its first move could get it stuck between a room and a corner of the
map, the rest of the map would go unfilled.

This behavior can be reproduced consistently by placing a room with
roompos.x = 5, roompos.y = 3, then starting the mazewalk with mm.x = 3
and mm.y = 5.

Adjust walkfrom so that it doesn't irrevocably commit to a direction
before recursing, which should solve this problem.  Fixes #56.