Fix: shopkeeper name inserted in bones level sound
The visit counter of a shopkeeper persists in a bones file; as a result,
bones which included a shop visited by a dearly departed hero would
insert the shopkeeper's name into the 'cursing shoplifters' message,
since the only criteria for that to happen was a positive visitct.

When entering a pre-visited shop in a bones file, the shopkeeper will
treat you like a returning customer if you were the most recent visitor
before the bones were saved (effectively, if they are your own bones),
and otherwise will not.  Use the same criteria when determining whether
to insert their name in the level sound message.