Track Groundhog Day as a holiday and allow woodchucks to generate then
Recent idea I had. Woodchucks are groundhogs, so giving them a nod on
Groundhog Day is a nifty little addition.

The generation change is made at the start of a game session, so
starting a game at 23:50 on Feb 1 and continuing to play it for hours
will see no woodchucks generate even though it's now Feb 2, and starting
a game at 23:50 on Feb 2 and continuing to play will see woodchucks
generate even though it's now Feb 3. Similar to how the moon's Luck
effect works.

Nothing is added regarding the folklore about a woodchuck seeing its
shadow -- doesn't really translate to a weatherless dungeon.