Add phoenixes
This is the second of two new monsters slated for version 7.0's
savebreaks. The implementation is mostly copied over from SpliceHack
(and the tile is too). Phoenixes are the new strongest member of class
B, and should appear primarily in the late midgame and late game. (Not
sure if this will make them appear in any special levels that just
specify B, intended for ravens and vampire bats, and now get tonally
wrong phoenixes instead. If so, that'd be considered a bug. After a
quick search it looks like the Wizard quest is the only place that
specifies this, and in my testing not even a level 30 wizard is managing
to generate any phoenixes.)

Minor other notable change is that revive_egg had to be externified, to
allow fire effects to revive a dead phoenix egg.