Optimization: only compute current holidays when needed
I want to expand use of holidays, but have been stopped by not wanting
to call it too much. Not that I seriously think it will harm
performance, but because it'd feel icky calling it once per turn or
whatever. Calling it every time a slime mold gets created already feels
like too much.

So this implements a system by which holidays only get recomputed when
there's been a day change to necessitate it. This is a little weird
since the game tracks days which begin at midnight and days which begin
at 6 PM, but overall it seems to work. If there hasn't been a change, it
just returns the holidays it computed last time.