Fix: exploding bag of holding gave experience too early
I realized that the sequence things happened in could award experience
and cause the player to gain a level and THEN subject them to flying bag
contents and the actual 6d6 explosion damage. That doesn't really make
sense, so this puts the experience gain at the end, only to be hit if
you survive the explosion.