Fix: impossible for rendering remembered erased engraving (stopgap)
The impossible is incorrect; it's possible for the game to try and
render an engraving glyph the hero only remembers about but which isn't
there anymore because a monster walked on and wiped it out.

Proper fix is to add separate glyphs for each engraving with their own
colors. This temporary fix just colors them brown on the basis that most
accidentally erased engravings will be dust.