Add engravings to the vanilla glyph system
This came about due to the slightly earlier bug fix for an impossible
when there was an erased-out-of-sight engraving. I realized that this
meant the shim code for displaying engravings in map_glyphinfo that I
added back was not going to be capable of correctly rendering a vanished
engraving, since it depended on looking up the state of the engraving on
that square. Whereas a glyph-based system would work, since then the
information about the engraving's appearance would be stored in the
"remembered" array.

I had previously been open to exploring trying this with engravings
because 1) unlike object materials it would actually be feasible to make
this small amount of tiles, 2) there was an ample example from the
devteam with how altar glyphs now work, and 3) it would be cool to have
different tiles for different engraving types. So I tried it out and
figured out how to make it work.

One casualty here: engr_to_glyph had to be a macro, and so it no longer
calls coord_hash. Now graffiti's color depends only on a simple sum of
its coordinates. Might wind up making the graffiti themed room look a
little predictable.