Adjust two tiles to prevent "gray mapping issue"
Apparently certain parts of the color palette are off-limits for
monsters (because monsters' tiles are also used for statues and not all
palette colors reduce to a given shade of gray). This adjusts Schliemann
and black mold so that they don't use such illegal colors.

While we're here, also augment the line that prints this warning message
out to actually tell you what tile and coordinate are being flagged,
since I needed those to figure out where the problem was in the first
place.