Squeaky boards squawk when hallucinating
I added the "You/foomonst step on a squeaky toy" as a hallu message for
this in the past, but it seemed a bit too on the nose and I was never
too much of a fan of it. This is a _little_ more streamlined (not too
much - hallucinatory messages still have to meet a minimum standard of
silliness), and still tells the player what note it is making.