Increase the prices of many magical tools
Considering their effects, it's a bit odd that some of these were so
cheap. For instance, a fire horn is similar to a wand of fire but with
better damage, and yet its base cost was a mere 50. Taming is a very
powerful effect, and yet a magic harp cost 50.

Most of the prices are quadrupled, though the magic whistle is increased
tenfold. In terms of price-identification of items this means that
drums, harps, and whistles can now be readily distinguished by price ID
where before it remained ambiguous.

Price of crystal balls was not changed, because I believe players are
already unlikely to buy them out of a shop in the first place.