Fix: impossible when monster attacked monster with c corpse
The flavorful hits patch made a critically wrong assumption that
everything that would be used to attack is a weapon. But monsters can
and do attack things with cockatrice corpses.

If attacking with a non-weapon, it should just default to "hit" as the
verb. The impossible is preserved but now only for the truly impossible
case where a weapon with no pierce/slash/blunt flags is used to attack.