Fix: is_open_air treated cloud terrain as unwalkable
Of course, clouds can't be walked on in real life, but video game logic
has a long tradition of walking on clouds. Beyond that, I don't want
players to read "trying to walk on open air is fatal" and assume it's
safe to e.g. jump onto a cloud and then find it's not safe. Also, it
occurred to me that there's a themed room with a single space of cloud,
or a region of cloud, in it and that shouldn't be treated as a precipice
either. There's a big room too if I remember right.

I think this bug came about just because I never realized IS_AIR also
includes clouds.

I don't have any intentions of using walkable clouds on levels that use
air, other than as decoration.