Fix: hero levelporting into level could end up in open air
Again, not obvious in Vlad's Cavern because that forces the hero to
arrive on a very limited patch of solid ground near the entry stairs.
Though it might be tempting to continue to let levelport regions control
this, it obviously shouldn't be a limiting factor on future level
designs.

This makes it so that open air spaces are now considered a "bad spot",
similar to spots occupied by monsters or most nonstandard terrain. (The
reason this bug happened in the first place is because air is explicitly
there as a "good" terrain type -- I guess for wizard mode levelporting
to the Plane of Air?) Even if you're flying or levitating, you won't
land on air -- this is consistent with how water and lava work.