Add magic platform terrain type
A magic platform is, essentially, a sturdy platform floating in open
air, and would only be expected to appear in conjunction with air.
This is designed for the new Valkyrie quest, where magic platforms will
be used to span a vast gulf of air. Like engravings, there are multiple
possible glyphs for the single S_magicplatform defsym; the non-default
ones will get used on the Valkyrie quest.

I considered not making a new terrain type and instead just adding
alternate glyphs for ROOM. However, I decided against that:
  1) It would require rearrangement of ROOM so that it's no longer in a
     cmap. No idea if this would cause serious issues.
  2) It would behave the same as any other ground, meaning the Valkyrie
     quest bridge would have weird behaviors like letting pits to get
     dug in it.

Magic platforms behave much the same as floor except they can't be dug
or engraved in. A possible future feature could be the ability to summon
temporary magic platforms which disappear after a while, using a terrain
timer like melting ice does.