Fix: destroying bear trap with pick-axe
Digging down with a pick axe on a bear trap you're trapped in will
destroy it.  This produced an impossible, "deleting trap containing
ammo?", because a bear trap uses its ammo field to store the bear trap
item which #untrapping it will produce.  Use deltrap_with_ammo instead,
to remove both the trap and its 'ammo'.