Implement redesigned Valkyrie locate level
This level has been planned for years. Tossing out the boring "cave of
Surtur" and replacing it with a much more dramatic setting: the rainbow
Bifrost bridge. In Norse legend, the bridge is prophesied to be broken
apart when Surtur's forces ride across it at the onset of Ragnarok.
Conveniently, the quest involves Surtur starting Ragnarok, so this works
very well. It also serves as a transition from the icy first half of the
quest to the fiery second half - which will actually be Muspellheim,
not a random cave.

The bridge is constructed of magic platforms - they behave normally, but
they use the six non-default magic platform glyphs to be actually
colored like a rainbow. One key challenge of this level is navigating
across the bridge, whether by jumping, levitating, or flying. (Or, I
suppose, by plunging off it to the lower filler level and tanking the
damage - the gaps in the bridge are air terrain, like Vlad's cavern.)

One relatively recent addition to this level is the frost giant leader
Hrymr, with a contingent of two other frost giants. I didn't want to
implement him as a new monster species, given that this is a bit of a
miniboss fight and not an actual boss fight, so instead he's just a
named frost giant with some good gear. No dialogue, either. This being a
later addition means I'm less attached to it: if it isn't well received,
he can be ditched.