Fix: double messages for created monsters appearing
This is a weird bug in that it's caused by a vanilla merge FAILING to
conflict. Basically, I implemented appearance messages for monsters in
create_critters a while back, then vanilla implemented a slightly more
general form of it the same way. The two implementations weren't in the
same function so git merge didn't notice.

I knew this wasn't the only change I had made to create_critters, so it
took a bit to sort out what parts of it were important to keep when most
of it was going away.