Killing Surtur restores the Valkyrie locate level bridge
Another long time planned feature of this quest redesign, and breaking
some new ground by having this be the first instance of a lua script
driving a terrain change well after the level is loaded.

Simple enough concept - having defeated Surtur and averted Ragnarok, the
damage he did to the locate level bridge is magically repaired. The
tricky bit is making sure this happens whether coming back to the level
after killing him, or killing him while ON the level. To that end, this
commit introduces a hopefully generalizable system in which a function
may be called that can do various components of such a level
transformation at different calling points.

Not totally happy with the code; it might require another use case or
two before settling into a final form. For now though it achieves the
goals of changing a level either when arriving to it or when killing the
nemesis on the level, without printing messages or resetting vision at
any wrong times.