Valk quest is completed not by getting artifact, but killing nemesis
One thing that has to change in order to give quests more varied
storylines is the idea of every quest being a mission to go get the
artifact, completed by presenting the artifact to the leader. That
doesn't allow enough flexibility to tell a different story for each of
the 13 roles.

One simple and obvious alternative goal for completing the quest is to
return to the quest leader having killed the nemesis. If you bring the
artifact back to the leader, they don't really care, because that's not
the main point anymore (and they need to have a special piece of quest
text to say in this situation, which is referred to by
"arti_but_not_neme". The Valkyrie quest adapts well to this goal since
the premise of Surtur not being a threat anymore if living but deprived
of the quest artifact is shaky.

This commit also puts in place some infrastructure for other quests
where returning an artifact isn't the goal.