Redesign Valkyrie quest upper filler level
Not an enormous change, but the lava vents persist, I threw in an
unfrozen lake for a change of scenery, and it's no longer a mines-style
filler. Plus the obligatory stairs at either end.

Also experimenting with having no "loot" lying around on the level but
instead carried by the filler monsters. So if you run around the giants
and don't fight them, no items for you. (This also makes more sense than
having random things lying around in the frozen wasteland.)