Change Valkyrie random quest enemies
This isn't changing the general quest system of having two species and
two classes and their fixed chances of spawning. The main threat is the
giants' incursion, and there are both fire and frost giants in Norse
mythology, so both deserve to get featured here. Meanwhile, fire ants
are good creatures to appear in Muspell but not really important to
feature all over the quest.

- 55% of random monsters will be fire giants
- 14% will be frost giants
- 17% will be random H
- 14% will be normal random monsters

I don't intend any files to specify random monsters created with the
level, so this will only affect spawning during play.