Redesign Valkyrie quest lower filler levels
Like the upper one, this is still a filler level. Not a huge change.
Instead of being in the bland "cave of Surtur", it's Muspelheim, but
both are the same fiery and lava-covered type of place. Just not a
mines-style floor; stairs are moved to opposite ends as per usual,
boulders scattered around (lots of giants live here), and a small chance
of a phoenix, which might be peaceful depending on alignment.