Fix: allow booleans to be passed to fields in des.door()
I was going to do this as a vanilla patch but then realized setting
locked/trapped on doors in a lua table is an xNetHack specific thing.

Basically, there's no reason why the level designer should be forced to
write locked=1 when there's a function that'll accept either ints or
bools as a boolean.