Replace the Orb of Fate with Sol Valtiva
This is part of the Valkyrie quest redesign, of course. One of the two
major changes to the story is that Lord Surtur didn't steal any artifact
from the home level. Instead, this is just his sword. You're under no
particular compulsion to do anything with it, other than that it's a
nice artifact to have. The goal of the quest is to kill him.

Sol Valtiva is attested in Norse mythology (though not strictly named as
such, there are just many references to Surtur's bright/flaming sword)
and implemented in dNetHack, though this implementation didn't really
draw on that. Its powers here deliberately stick closer to what the Orb
of Fate provided. In diff form:
- warning when carried
- usable as crystal ball
- acts as a luckstone
+ drops from 100 to 93 weight
+ +d3 to hit, +d5 fire damage (this isn't supposed to be Fire Brand 2.0)
+ usable as a two-handed sword
unchanged: half physical damage, half spell damage, invoke for levelport

Balancing the powerset is tricky, especially since this is replacing a
non-weapon artifact with a weapon: I do not want it to be so useless
that it's not worth carrying around, but I don't want it to be so
dominating that every Valkyrie uses it as their main weapon for the rest
of the game. I hope the combination of good effects you don't need with
the disadvantages of a two-handed sword that is only somewhat better
than a regular one combat-wise hit this middle ground.

There were also a number of options available for giving additional cool
powers with the "bright" aspect of the sword, such as creating lit areas
on hits or blinding monsters.

One subtle adjustment with important ramifications: this artifact is now
chaotic (because, obviously, it's Surtur's.) Since it's still
implemented as the Valkyrie quest artifact, they won't have any problem
with this, but it significantly changes artifact wishing for other
roles. In general, I would imagine many wishes for the Master Key of
Thievery would switch over to this because it provides both half
damages (though the Key is light and this is heavy). In the same stroke,
it also reduces the neutral artifact pool. In general, people consider
the neutral pool to be stuffed with good artifacts while the chaotic
pool is crappy, so hopefully this evens it out a bit.