Redesign Valkyrie goal level
There was nothing super wrong with the existing goal level, so this just
tweaks the design a bit to better reflect how the story will play out.
Overall still keeps the lava aesthetic. Surtur now carries Sol Valtiva
(as in the redesigned Arc quest, he's not just standing on top of it)
and a wand of fire. The fortress complex is larger, containing half of
the old one. (Still has a drawbridge, but in a different location.) Also
plenty of boulders lying around like in the rest of the quest, and
appropriately fiery Muspelheim denizens, including fire ants.