Revise Valkyrie quest text
Locking in the details of the revised story. Some text worked for the
new text and didn't really need to be changed; other bits had to be
updated to reflect the new levels or the goal of killing Surtur rather
than getting a stolen artifact.

One bit in particular that was satisfying to change was Lord Surtur's
death message, because the old one where he appears to die of a pounding
headache was just silly.