Disallow wishing for enchantment on dragon scales
This has been a loophole in the dtsund-dsm dragon scale implementation
that I was initially okay with, but of course as soon as it was
discovered it quickly became established strategy to wish up a +3 or +5
set of dragon scales as an endgame cloak, which defeats much of the
purpose of encouraging found objects rather than using wishes to build
an endgame set of armor.

It's not possible to give dragon scales an enchantment by other means,
so close the loophole and make it impossible to give dragon scales an
enchantment at all.