Fix: stop shatterproof glass items from shattering in several cases
1. Glass items in a chest being kicked
2. Weapon breaking because of forcing the lock on a box
3. Shattering an enemy's weapon in melee
4. Breaking a lance on a bad joust

Note that in most of these cases, non-glass erodeproof items will still
shatter. The point of this fix isn't to make all erodeproofed items
immune to these shattering mechanics, it's to make it so that there
aren't items which are labeled "shatterproof" by the game but break
anyway. Also doesn't affect glass ammo mulching, which isn't necessarily
breakage; could also be loss.