Allow artifacts created by naming to keep their original material
This averts the flavor oddity of having an elven dagger of whatever
non-base material (wood, mithril, etc...) that you then name as Sting,
and suddenly it's made of copper.

For players that like silver artifacts, this means you can indeed get a
silver Sting or Orcrist provided you can find a silver base weapon for
it.