Fix: items dropped from various sources hovered in midair
Reported by several sources, about killing a dragon or stoning a monster
in Vlad's Cavern and the dragon scales or the statue would just sit
there hanging in midair. This potentially extended to anything that
called mkobj_at or mksobj_at, directly or indirectly. (Those just
place_object without calling flooreffects.)

The fundamental problem is that there are parts of the code that need to
create an object, modify it in some way (e.g. hardcoded death drops from
certain monster types, such as a unicorn horn being degraded if the
slain unicorn was revived), and only after such modifications can it be
potentially dropped into midair and lost. So we can't make mksobj_at
call flooreffects, because it needs to be guaranteed to return an
object.

Instead, I exhaustively went through all the possible calls of these
(and indirect calls, such as mkcorpstat) looking for cases where the
created object could end up on top of air during normal gameplay. These
now call a new function obj_drops_at (which takes care of the
flooreffects and so on) at an appropriate point when the object is able
to be destroyed without adverse effect.

There are various cases not touched:
- Level creation in general. Note that there CAN be stuff generated on
  air if some themed room designer creates a room that contains air and
  allows it to be either "ordinary" type or filled = 1, so I added a
  warning at the top of themerms.lua not to do that.
- Where an object is being dropped on a known terrain type that isn't
  going to do anything to it (gem being found in fountain, headstone
  breaking into rocks, etc.)
- Object creation in Lua; if Lua code places an item on open air it's a
  bug in that Lua code. I don't currently conceive of any script that
  would *intentionally* dump objects into open air.
- Creation of the player's corpse/statue because bones can't be saved
  when the player died over air; clarified that the reason for
  disallowing bones here is not to avoid hovering items but because it
  doesn't make for interesting bones. (But also, it avoids hovering
  items.)