Foocubi gain a level when draining one from the player
Because the level of a foocubus impacts its odds of giving the hero a
negative result, this change has the effect of slightly tipping the
balance of repeated interaction with the same foocubus in the foocubus's
favor. Over time, the foocubus will accumulate levels (even if the hero
is using restore ability or full healing or just killing monsters to
regain levels they lose) and make subsequent good outcomes harder to
gain.

If the hero is drain resistant or protected from losing the level by
enchanted bone armor intercepting the drain, the foocubus doesn't gain
a level, because they didn't successfully take one away.