Don't hide magic portals
This is a small but significant change. Main reason for it is to avoid
the minor tedium of playing Hunt For The Magic Square on the quest
portal level. But beyond that, there's no good reason a magic portal
would be invisible by default. I at least envision it as something
obviously magical and portal-like, rather than a patch of air that looks
the same as any other until you step onto it.