Gnomes outside the Mines have a chance of generating with a few gems
This is a bit of flavor (gnomes that have ventured out of the Mines are
probably there to trade) and also a bit of a nice side effect for
players encountering gnomes outside their normal habitat. This does not
apply in the Mines because there are plenty of gem-carrying gnomes there
already.