Give quantum mechanics physicist names when farlooked
This is a subset of the "Named Monsters v1.0" 3.4.3 patch. Apparently
that in turn was an outgrowth of a feature in vanilla 3.0 where quantum
mechanics were named.

I dropped a couple names for various reasons, and got rid of the system
where a quantum mechanic's gender would determine its pool of names -
since the pool of female names is very noticeably smaller.

Also, this uses the same system as coyotes - whereas the patch injected
it into most naming functions so you would see the scientist name in
combat, etc. This only shows it when you farlook, and I think the
subtlety greatly improves the feature.