Reset cknown when thiefstone changes box contents
While playing today, I picked up what turned out to be a thiefstone that
grabbed something from my pack and vanished.  I remembered that there
was a chest across the level, so I farlooked it and could see that the
description changed from "an empty unlocked chest" to "an unlocked chest
containing 2 items".  It seemed odd to me that the hero could detect
that change from across the level.

Change this: when a thiefstone deposits something into its stash
container, reset container->cknown.  There will still be a change
visible in farlook, but this approach is is more consistent with things
that modify container contents outside of direct #looting (e.g. monster
rummaging).