Fix: reversed test in do_symset
In vanilla these tests are `if (rogueflag ? sl->primary : sl->rogue)`;
in other words, if looking for a Rogue symset, exclude anything marked
as being eligible only for the primary symset, and vice-versa.  When the
Rogue symset was ripped out, apparently the wrong part of the ternary
was preserved, so it was excluding anything marked as primary-only
instead of anything marked rogue-only.  Flip it back around.