Fix: becoming unpunished over open air
More B&C and open air interactions: if the hero is punished, starts
levitating over a chasm of open air, then gets unpunished via a scroll
of remove curse or similar, her heavy iron ball becomes detached and
therefore should fall down to the lower level.  Instead, it stayed where
it was, floating on the air.  If the ball is sitting on an air spot when
unpunish(read.c) is run, call flooreffects to drop it if necessary.