Fix: potential use-after-free in obj_aireffects
An object could be deleted in obj_aireffects (if it broke or fell into
a bottomless pit), and then subsequently used for xname, etc.  Save the
object's name before deleting it.  Also clear next_boulder in
obj_aireffects rather than before, so the use of "the next boulder" will
match pushing a stack of boulders into lava or water.