Allow drawbridges to stay jammed
Mentioned in 1654735 and a comment in do_play_instrument(music.c) as
something which would be a nice future improvement if there were more
bits available in flags/drawbridgemask.  Turns out the DB_UNDER states
are all mutually exclusive and never overlap, so a bit for the jam can
be saved by rearranging those.