Remove level-change reset_glyphmap call
As far as I know the only reason for calling reset_glyphmap upon every
level change is in case the hero is entering or exiting the rogue level
and the rogue symset/colors need to be switched with the normal
symset/colors.  Since xnh has removed the rogue level, this shouldn't be
necessary, and removing it will reduce the amount of time spent on
iterating through the increased number of glyphs.