Improve(?) on ceiling-required trap generation
This is arguably a better approach than the previous one, in that it
allows for rock traps to generate randomly indoors (or under some other
kind of ceiling) even on levels that are flagged 'outdoors'.  However,
it's a bit wasteful since any outdoors level without 'indoor' areas will
spend a bunch of cycles trying to place the trap before finally giving
up.  Maybe not that big a deal in the big picture though.