Add wizmode option to highlight ceilings
I wouldn't include this in the actual release/master branch, since it
sloppily relies on overwriting the color and piggybacking off the
inverse video effect of object piles, but it's useful for testing
purposes.  If this were to be a permanent addition to the game I'd add
some new flag like MG_CEILING that could be used for this purpose, but I
doubt it's worth it.  I'm including it in this branch just so that
others can use it for testing if desired, because it's convenient.