When a monster spawns with a weapon, it wields it immediately
Often times the player could take advantage of this "free turn" to
attack a monster that just got created, since it would spend that turn
on wielding the weapon first thing.

The way this is implemented, it only applies to monsters created after
level creation time and who are not created to be asleep; I think this
issue didn't affect monsters generated with the level who are either up
and moving about and have time to wield their weapon, or asleep (e.g. a
barracks) in which case a free turn to attack may be justified.

The addition in create_particular_creation is redundant in terms of that
function, but prevents wizard-mode creation of a sleeping monster
wielding its weapon (but remaining asleep).